Move `EncryptedChoice` to `app` module
