+++ src/group/ristretto.rs
@@ -118,7 +118,7 @@ mod tests {
     use rand::thread_rng;
 
     use super::*;
-    use crate::{group::Curve25519Subgroup, DiscreteLogTable};
+    use crate::{app::EncryptedChoice, group::Curve25519Subgroup, DiscreteLogTable};
 
     type SecretKey = crate::SecretKey<Ristretto>;
     type Keypair = crate::Keypair<Ristretto>;
@@ -137,8 +137,8 @@ mod tests {
     fn encrypt_choice() {
         let mut rng = thread_rng();
         let keypair = Keypair::generate(&mut rng);
-        let encrypted_choice = keypair.public().encrypt_choice(5, 3, &mut rng);
-        let variant_ciphertexts = keypair.public().verify_choice(&encrypted_choice).unwrap();
+        let encrypted_choice = EncryptedChoice::new(5, 3, keypair.public(), &mut rng);
+        let variant_ciphertexts = encrypted_choice.verify(keypair.public()).unwrap();
 
         let lookup_table = DiscreteLogTable::new(0..=1);
         for (i, &variant) in variant_ciphertexts.iter().enumerate() {
