+++ tests/integration/basic.rs
@@ -5,7 +5,7 @@ use rand::{thread_rng, Rng};
 use std::collections::HashMap;
 
 use crate::assert_ct_eq;
-use elastic_elgamal::{group::Group, Keypair, LogEqualityProof, RingProof};
+use elastic_elgamal::{app::EncryptedChoice, group::Group, Keypair, LogEqualityProof, RingProof};
 
 fn test_encryption_roundtrip<G: Group>() {
     let mut rng = thread_rng();
@@ -143,8 +143,8 @@ fn test_encrypted_choice_works<G: Group>() {
     let mut rng = thread_rng();
     let keypair = Keypair::<G>::generate(&mut rng);
 
-    let choice = keypair.public().encrypt_choice(5, 2, &mut rng);
-    let variants = keypair.public().verify_choice(&choice).unwrap();
+    let choice = EncryptedChoice::new(5, 2, keypair.public(), &mut rng);
+    let variants = choice.verify(keypair.public()).unwrap();
     assert_eq!(variants.len(), 5);
     for (i, &variant) in variants.iter().enumerate() {
         let expected_plaintext = if i == 2 {
