+++ tests/snapshots.rs
@@ -5,6 +5,7 @@ use rand_chacha::ChaChaRng;
 use rand_core::SeedableRng;
 
 use elastic_elgamal::{
+    app::EncryptedChoice,
     group::{Generic, Group, Ristretto},
     Keypair, RangeDecomposition,
 };
@@ -74,7 +75,7 @@ fn test_encrypted_choice_snapshot<G: Group + Named>() {
     let mut rng = ChaChaRng::seed_from_u64(12345);
     let (public_key, _) = Keypair::<G>::generate(&mut rng).into_tuple();
 
-    let choice = public_key.encrypt_choice(5, 3, &mut rng);
+    let choice = EncryptedChoice::new(5, 3, &public_key, &mut rng);
     let full_name = format!("encrypted-choice-{}", G::NAME);
     assert_yaml_snapshot!(full_name, choice);
 }
